Move banner UI component from general compose package to payments package
