+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/MainSettingsFragment.kt
@@ -35,8 +35,8 @@ import com.woocommerce.android.extensions.show
 import com.woocommerce.android.model.FeatureAnnouncement
 import com.woocommerce.android.support.HelpActivity
 import com.woocommerce.android.support.HelpActivity.Origin
-import com.woocommerce.android.ui.compose.component.banner.SettingsBannerDismissDialog
-import com.woocommerce.android.ui.compose.component.banner.SettingsScreenBanner
+import com.woocommerce.android.ui.payments.banner.SettingsBannerDismissDialog
+import com.woocommerce.android.ui.payments.banner.SettingsScreenBanner
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderFlowParam
 import com.woocommerce.android.util.AnalyticsUtils
