+++ WooCommerce/src/main/res/layout/fragment_settings_main.xml
@@ -18,7 +18,8 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"/>
 
-        <View style="@style/Woo.Divider" />
+        <View style="@style/Woo.Divider"
+            android:background="@color/woo_black_90_alpha_020"/>
 
         <!--
             Help & support
