+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/prefs/MainPresenterTest.kt
@@ -3,7 +3,7 @@ package com.woocommerce.android.ui.prefs
 import com.woocommerce.android.AppUrls
 import com.woocommerce.android.analytics.AnalyticsTracker.Companion.KEY_BANNER_PAYMENTS
 import com.woocommerce.android.tools.SelectedSite
-import com.woocommerce.android.ui.compose.component.banner.BannerDisplayEligibilityChecker
+import com.woocommerce.android.ui.payments.banner.BannerDisplayEligibilityChecker
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
